エディター部の独立化
