+++ AOJ_App/AOJ_App/MainWindow.xaml.cs
@@ -26,40 +26,11 @@ namespace AOJ_App
     /// </summary>
     public partial class MainWindow : Window
     {
-        private CompletionWindow completewind;
         public MainWindow()
         {
             InitializeComponent();
-            editor.TextArea.TextEntered += TextArea_TextEntered;
-            editor.TextArea.TextEntering += TextArea_TextEntering;
-            
-            
+            var Editor = new codeeditor(editor);
         }
 
-        private void TextArea_TextEntering(object sender, TextCompositionEventArgs e)
-        {
-            if(e.Text.Length > 0 && completewind != null)
-            {
-                if (!char.IsLetterOrDigit(e.Text[0]))
-                {
-                   completewind.CompletionList.RequestInsertion(e);
-                }
-            }
-        }
-
-        private async void TextArea_TextEntered(object sender, TextCompositionEventArgs e)
-        {
-            if(e.Text == ".")
-            {
-                completewind = new CompletionWindow(editor.TextArea);
-                IList<ICompletionData> data = completewind.CompletionList.CompletionData;
-                data.Add(new CompletionData("test1"));
-                completewind.Show();
-                completewind.Closed += delegate
-                {
-                    completewind = null;
-                };
-            }
-        }
     }
 }
